requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.3
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
pylint==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.4
razorpay==1.2.0
reportlab==3.5.56
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
twilio==6.51.1
urllib3==1.26.2
virtualenv==20.2.2
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
xhtml2pdf==0.2.5
django-admin-interface
python-dotenv
